fix(quality): yaml-safe frontmatter write + git_status rollback (review feedback PR #1153)
